backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["patrickezway <patrick@ezway-technology.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.20.0"
fastapi = {extras = ["all"], version = "^0.89.1"}
fastapi-users = {extras = ["sqlalchemy"], version = "^10.2.1"}
aiosqlite = "^0.18.0"
databases = {extras = ["sqlite"], version = "^0.7.0"}
python-dotenv = "^0.21.0"
python-multipart = "^0.0.5"
jinja2 = "^3.1.2"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^1.4.46"
mangum = "^0.17.0"
asgiref = "^3.6.0"
click = "^8.1.3"
dnspython = "^2.3.0"
h11 = "^0.14.0"
pymongo = "^4.3.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"